Move Syncable::Model into Syncable concern
